+++ packages/schema-blocks/src/core/blocks/index.ts
@@ -1,0 +1,4 @@
+import BlockInstruction from "./BlockInstruction";
+
+export { BlockInstruction };
+
