+++ packages/schema-blocks/src/core/index.ts
@@ -1,0 +1,2 @@
+export { BlockInstruction } from "./blocks";
+export { BlockValidation, BlockValidationResult } from "./validation";
