+++ packages/schema-blocks/src/instructions/blocks/index.ts
@@ -14,3 +14,7 @@ import "./TextInput";
 import "./VariableTagRichText";
 import "./Variation";
 import "./VariationPicker";
+
+import VariableTagRichText from "./VariableTagRichText";
+
+export { VariableTagRichText };
