+++ packages/schema-blocks/src/instructions/index.ts
@@ -1,2 +1,2 @@
-import "./blocks";
-import "./schema";
+export * from "./blocks";
+export * from  "./schema";
